refactor(gateway): extract external-agent runtime handlers (#2969)
