+++ crates/tau-gateway/src/gateway_openresponses.rs
@@ -37,9 +37,7 @@ use tau_multi_channel::multi_channel_lifecycle::{
 };
 use tau_runtime::{
     inspect_runtime_heartbeat, start_runtime_heartbeat_scheduler, ExternalCodingAgentBridge,
-    ExternalCodingAgentBridgeConfig, ExternalCodingAgentBridgeError,
-    ExternalCodingAgentSessionSnapshot, ExternalCodingAgentSessionStatus,
-    RuntimeHeartbeatSchedulerConfig, TransportHealthSnapshot,
+    ExternalCodingAgentBridgeConfig, RuntimeHeartbeatSchedulerConfig, TransportHealthSnapshot,
 };
 use tau_session::SessionStore;
 use tokio::net::TcpListener;
@@ -56,6 +54,7 @@ mod dashboard_runtime;
 mod dashboard_shell_page;
 mod dashboard_status;
 mod deploy_runtime;
+mod external_agent_runtime;
 mod jobs_runtime;
 mod memory_runtime;
 mod multi_channel_status;
@@ -97,6 +96,13 @@ use dashboard_status::{
     collect_tau_ops_dashboard_command_center_snapshot, GatewayDashboardActionRequest,
 };
 use deploy_runtime::{handle_gateway_agent_stop, handle_gateway_deploy};
+use external_agent_runtime::{
+    handle_external_coding_agent_open_session, handle_external_coding_agent_reap,
+    handle_external_coding_agent_session_close, handle_external_coding_agent_session_detail,
+    handle_external_coding_agent_session_followup,
+    handle_external_coding_agent_session_followups_drain,
+    handle_external_coding_agent_session_progress, handle_external_coding_agent_session_stream,
+};
 use jobs_runtime::{handle_gateway_job_cancel, handle_gateway_jobs_list};
 use memory_runtime::{
     gateway_memory_store, gateway_memory_stores_root, handle_api_memories_graph,
@@ -131,10 +137,7 @@ use tools_runtime::{handle_gateway_tools_inventory, handle_gateway_tools_stats};
 use training_runtime::{handle_gateway_training_config_patch, handle_gateway_training_rollouts};
 use types::{
     GatewayAuthBootstrapResponse, GatewayAuthSessionRequest, GatewayAuthSessionResponse,
-    GatewayChannelLifecycleRequest, GatewayConfigPatchRequest,
-    GatewayExternalCodingAgentFollowupsDrainRequest, GatewayExternalCodingAgentMessageRequest,
-    GatewayExternalCodingAgentReapRequest, GatewayExternalCodingAgentSessionOpenRequest,
-    GatewayExternalCodingAgentStreamQuery, GatewayMemoryEntryDeleteRequest,
+    GatewayChannelLifecycleRequest, GatewayConfigPatchRequest, GatewayMemoryEntryDeleteRequest,
     GatewayMemoryEntryUpsertRequest, GatewayMemoryGraphEdge, GatewayMemoryGraphFilterSummary,
     GatewayMemoryGraphNode, GatewayMemoryGraphQuery, GatewayMemoryGraphResponse,
     GatewayMemoryReadQuery, GatewayMemoryUpdateRequest, GatewaySafetyPolicyUpdateRequest,
@@ -1523,353 +1526,6 @@ fn collect_gateway_events_status_report(gateway_state_dir: &Path) -> GatewayEven
     }
 }
 
-async fn handle_external_coding_agent_open_session(
-    State(state): State<Arc<GatewayOpenResponsesServerState>>,
-    headers: HeaderMap,
-    body: Bytes,
-) -> Response {
-    if let Err(error) = authorize_and_enforce_gateway_limits(&state, &headers) {
-        return error.into_response();
-    }
-    if let Err(error) = validate_gateway_request_body_size(&state, &body) {
-        return error.into_response();
-    }
-    let request =
-        match parse_gateway_json_body::<GatewayExternalCodingAgentSessionOpenRequest>(&body) {
-            Ok(request) => request,
-            Err(error) => return error.into_response(),
-        };
-    state
-        .external_coding_agent_bridge
-        .reap_inactive_sessions(current_unix_timestamp_ms());
-    let snapshot = match state
-        .external_coding_agent_bridge
-        .open_or_reuse_session(request.workspace_id.as_str())
-    {
-        Ok(snapshot) => snapshot,
-        Err(error) => return map_external_coding_agent_bridge_error(error).into_response(),
-    };
-    record_cortex_external_session_opened(
-        &state.config.state_dir,
-        snapshot.session_id.as_str(),
-        snapshot.workspace_id.as_str(),
-        external_coding_agent_status_label(snapshot.status),
-    );
-    (
-        StatusCode::OK,
-        Json(json!({
-            "session": external_coding_agent_session_json(&snapshot),
-        })),
-    )
-        .into_response()
-}
-
-async fn handle_external_coding_agent_session_detail(
-    State(state): State<Arc<GatewayOpenResponsesServerState>>,
-    headers: HeaderMap,
-    AxumPath(session_id): AxumPath<String>,
-) -> Response {
-    if let Err(error) = authorize_and_enforce_gateway_limits(&state, &headers) {
-        return error.into_response();
-    }
-    state
-        .external_coding_agent_bridge
-        .reap_inactive_sessions(current_unix_timestamp_ms());
-    let Some(snapshot) = state
-        .external_coding_agent_bridge
-        .snapshot(session_id.as_str())
-    else {
-        return OpenResponsesApiError::not_found(
-            "external_coding_agent_session_not_found",
-            format!("session '{session_id}' was not found"),
-        )
-        .into_response();
-    };
-    (
-        StatusCode::OK,
-        Json(json!({ "session": external_coding_agent_session_json(&snapshot) })),
-    )
-        .into_response()
-}
-
-async fn handle_external_coding_agent_session_progress(
-    State(state): State<Arc<GatewayOpenResponsesServerState>>,
-    headers: HeaderMap,
-    AxumPath(session_id): AxumPath<String>,
-    body: Bytes,
-) -> Response {
-    if let Err(error) = authorize_and_enforce_gateway_limits(&state, &headers) {
-        return error.into_response();
-    }
-    if let Err(error) = validate_gateway_request_body_size(&state, &body) {
-        return error.into_response();
-    }
-    let request = match parse_gateway_json_body::<GatewayExternalCodingAgentMessageRequest>(&body) {
-        Ok(request) => request,
-        Err(error) => return error.into_response(),
-    };
-    let event = match state
-        .external_coding_agent_bridge
-        .append_progress(session_id.as_str(), request.message.as_str())
-    {
-        Ok(event) => event,
-        Err(error) => return map_external_coding_agent_bridge_error(error).into_response(),
-    };
-    record_cortex_external_progress_event(
-        &state.config.state_dir,
-        session_id.as_str(),
-        event.sequence_id,
-        event.message.as_str(),
-    );
-    let session = state
-        .external_coding_agent_bridge
-        .snapshot(session_id.as_str())
-        .map(|snapshot| external_coding_agent_session_json(&snapshot))
-        .unwrap_or_else(|| Value::Null);
-    (
-        StatusCode::OK,
-        Json(json!({
-            "event": external_coding_agent_event_json(&event),
-            "session": session,
-        })),
-    )
-        .into_response()
-}
-
-async fn handle_external_coding_agent_session_followup(
-    State(state): State<Arc<GatewayOpenResponsesServerState>>,
-    headers: HeaderMap,
-    AxumPath(session_id): AxumPath<String>,
-    body: Bytes,
-) -> Response {
-    if let Err(error) = authorize_and_enforce_gateway_limits(&state, &headers) {
-        return error.into_response();
-    }
-    if let Err(error) = validate_gateway_request_body_size(&state, &body) {
-        return error.into_response();
-    }
-    let request = match parse_gateway_json_body::<GatewayExternalCodingAgentMessageRequest>(&body) {
-        Ok(request) => request,
-        Err(error) => return error.into_response(),
-    };
-    let event = match state
-        .external_coding_agent_bridge
-        .queue_followup(session_id.as_str(), request.message.as_str())
-    {
-        Ok(event) => event,
-        Err(error) => return map_external_coding_agent_bridge_error(error).into_response(),
-    };
-    record_cortex_external_followup_event(
-        &state.config.state_dir,
-        session_id.as_str(),
-        event.sequence_id,
-        event.message.as_str(),
-    );
-    let session = state
-        .external_coding_agent_bridge
-        .snapshot(session_id.as_str())
-        .map(|snapshot| external_coding_agent_session_json(&snapshot))
-        .unwrap_or_else(|| Value::Null);
-    (
-        StatusCode::OK,
-        Json(json!({
-            "event": external_coding_agent_event_json(&event),
-            "session": session,
-        })),
-    )
-        .into_response()
-}
-
-async fn handle_external_coding_agent_session_followups_drain(
-    State(state): State<Arc<GatewayOpenResponsesServerState>>,
-    headers: HeaderMap,
-    AxumPath(session_id): AxumPath<String>,
-    body: Bytes,
-) -> Response {
-    if let Err(error) = authorize_and_enforce_gateway_limits(&state, &headers) {
-        return error.into_response();
-    }
-    if let Err(error) = validate_gateway_request_body_size(&state, &body) {
-        return error.into_response();
-    }
-    let request = if body.is_empty() {
-        GatewayExternalCodingAgentFollowupsDrainRequest::default()
-    } else {
-        match parse_gateway_json_body::<GatewayExternalCodingAgentFollowupsDrainRequest>(&body) {
-            Ok(request) => request,
-            Err(error) => return error.into_response(),
-        }
-    };
-    let limit = request.limit.unwrap_or(64).max(1);
-    let followups = match state
-        .external_coding_agent_bridge
-        .take_followups(session_id.as_str(), limit)
-    {
-        Ok(followups) => followups,
-        Err(error) => return map_external_coding_agent_bridge_error(error).into_response(),
-    };
-    let session = state
-        .external_coding_agent_bridge
-        .snapshot(session_id.as_str())
-        .map(|snapshot| external_coding_agent_session_json(&snapshot))
-        .unwrap_or_else(|| Value::Null);
-    (
-        StatusCode::OK,
-        Json(json!({
-            "session_id": session_id,
-            "drained_count": followups.len(),
-            "followups": followups,
-            "session": session,
-        })),
-    )
-        .into_response()
-}
-
-async fn handle_external_coding_agent_session_stream(
-    State(state): State<Arc<GatewayOpenResponsesServerState>>,
-    headers: HeaderMap,
-    AxumPath(session_id): AxumPath<String>,
-    Query(query): Query<GatewayExternalCodingAgentStreamQuery>,
-) -> Response {
-    if let Err(error) = authorize_and_enforce_gateway_limits(&state, &headers) {
-        return error.into_response();
-    }
-    state
-        .external_coding_agent_bridge
-        .reap_inactive_sessions(current_unix_timestamp_ms());
-    let Some(snapshot) = state
-        .external_coding_agent_bridge
-        .snapshot(session_id.as_str())
-    else {
-        return OpenResponsesApiError::not_found(
-            "external_coding_agent_session_not_found",
-            format!("session '{session_id}' was not found"),
-        )
-        .into_response();
-    };
-    let replay_from_header = headers
-        .get("last-event-id")
-        .and_then(|value| value.to_str().ok())
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .and_then(|value| value.parse::<u64>().ok());
-    let after_sequence_id = query.after_sequence_id.or(replay_from_header);
-    let limit = query
-        .limit
-        .unwrap_or(
-            state
-                .config
-                .external_coding_agent_bridge
-                .max_events_per_session,
-        )
-        .max(1);
-    let events = match state.external_coding_agent_bridge.poll_events(
-        session_id.as_str(),
-        after_sequence_id,
-        limit,
-    ) {
-        Ok(events) => events,
-        Err(error) => return map_external_coding_agent_bridge_error(error).into_response(),
-    };
-
-    let (tx, rx) = mpsc::unbounded_channel::<SseFrame>();
-    let _ = tx.send(SseFrame::Json {
-        event: "external_coding_agent.snapshot",
-        payload: json!({
-            "session": external_coding_agent_session_json(&snapshot),
-            "replay_after_sequence_id": after_sequence_id,
-        }),
-    });
-    for event in events {
-        let _ = tx.send(SseFrame::Json {
-            event: "external_coding_agent.progress",
-            payload: external_coding_agent_event_json(&event),
-        });
-    }
-    let _ = tx.send(SseFrame::Done);
-    drop(tx);
-
-    let stream =
-        UnboundedReceiverStream::new(rx).map(|frame| Ok::<Event, Infallible>(frame.into_event()));
-    Sse::new(stream)
-        .keep_alive(KeepAlive::default())
-        .into_response()
-}
-
-async fn handle_external_coding_agent_session_close(
-    State(state): State<Arc<GatewayOpenResponsesServerState>>,
-    headers: HeaderMap,
-    AxumPath(session_id): AxumPath<String>,
-) -> Response {
-    if let Err(error) = authorize_and_enforce_gateway_limits(&state, &headers) {
-        return error.into_response();
-    }
-    let snapshot = match state
-        .external_coding_agent_bridge
-        .close_session(session_id.as_str())
-    {
-        Ok(snapshot) => snapshot,
-        Err(error) => return map_external_coding_agent_bridge_error(error).into_response(),
-    };
-    record_cortex_external_session_closed(
-        &state.config.state_dir,
-        snapshot.session_id.as_str(),
-        snapshot.workspace_id.as_str(),
-        external_coding_agent_status_label(snapshot.status),
-    );
-    (
-        StatusCode::OK,
-        Json(json!({
-            "session": external_coding_agent_session_json(&snapshot),
-        })),
-    )
-        .into_response()
-}
-
-async fn handle_external_coding_agent_reap(
-    State(state): State<Arc<GatewayOpenResponsesServerState>>,
-    headers: HeaderMap,
-    body: Bytes,
-) -> Response {
-    if let Err(error) = authorize_and_enforce_gateway_limits(&state, &headers) {
-        return error.into_response();
-    }
-    if let Err(error) = validate_gateway_request_body_size(&state, &body) {
-        return error.into_response();
-    }
-    let request = if body.is_empty() {
-        GatewayExternalCodingAgentReapRequest::default()
-    } else {
-        match parse_gateway_json_body::<GatewayExternalCodingAgentReapRequest>(&body) {
-            Ok(request) => request,
-            Err(error) => return error.into_response(),
-        }
-    };
-    let now_unix_ms = request
-        .now_unix_ms
-        .unwrap_or_else(current_unix_timestamp_ms);
-    let sessions = state
-        .external_coding_agent_bridge
-        .reap_inactive_sessions(now_unix_ms)
-        .into_iter()
-        .map(|snapshot| external_coding_agent_session_json(&snapshot))
-        .collect::<Vec<_>>();
-    (
-        StatusCode::OK,
-        Json(json!({
-            "reaped_count": sessions.len(),
-            "sessions": sessions,
-            "runtime": {
-                "active_sessions": state.external_coding_agent_bridge.active_session_count(),
-                "inactivity_timeout_ms": state.config.external_coding_agent_bridge.inactivity_timeout_ms,
-                "max_active_sessions": state.config.external_coding_agent_bridge.max_active_sessions,
-                "max_events_per_session": state.config.external_coding_agent_bridge.max_events_per_session,
-            }
-        })),
-    )
-        .into_response()
-}
-
 async fn handle_openresponses(
     State(state): State<Arc<GatewayOpenResponsesServerState>>,
     headers: HeaderMap,
@@ -2742,81 +2398,6 @@ fn parse_gateway_json_body<T: DeserializeOwned>(body: &Bytes) -> Result<T, OpenR
             "malformed_json",
             format!("failed to parse request body: {error}"),
         )
-    })
-}
-
-fn map_external_coding_agent_bridge_error(
-    error: ExternalCodingAgentBridgeError,
-) -> OpenResponsesApiError {
-    match error {
-        ExternalCodingAgentBridgeError::InvalidWorkspaceId => OpenResponsesApiError::bad_request(
-            "invalid_workspace_id",
-            "workspace_id must be non-empty",
-        ),
-        ExternalCodingAgentBridgeError::InvalidMessage => {
-            OpenResponsesApiError::bad_request("invalid_message", "message must be non-empty")
-        }
-        ExternalCodingAgentBridgeError::InvalidSubprocessConfig(message) => {
-            OpenResponsesApiError::internal(format!(
-                "external coding-agent subprocess configuration is invalid: {message}"
-            ))
-        }
-        ExternalCodingAgentBridgeError::SessionNotFound(session_id) => {
-            OpenResponsesApiError::not_found(
-                "external_coding_agent_session_not_found",
-                format!("session '{session_id}' was not found"),
-            )
-        }
-        ExternalCodingAgentBridgeError::SessionLimitReached { limit } => {
-            OpenResponsesApiError::new(
-                StatusCode::CONFLICT,
-                "external_coding_agent_session_limit_reached",
-                format!("max active sessions limit reached ({limit})"),
-            )
-        }
-        ExternalCodingAgentBridgeError::SubprocessSpawnFailed {
-            workspace_id,
-            error,
-        } => OpenResponsesApiError::gateway_failure(format!(
-            "failed to start external coding-agent worker for workspace '{workspace_id}': {error}"
-        )),
-        ExternalCodingAgentBridgeError::SubprocessIoError { session_id, error } => {
-            OpenResponsesApiError::gateway_failure(format!(
-                "external coding-agent worker I/O failed for session '{session_id}': {error}"
-            ))
-        }
-    }
-}
-
-fn external_coding_agent_status_label(status: ExternalCodingAgentSessionStatus) -> &'static str {
-    match status {
-        ExternalCodingAgentSessionStatus::Running => "running",
-        ExternalCodingAgentSessionStatus::Completed => "completed",
-        ExternalCodingAgentSessionStatus::Failed => "failed",
-        ExternalCodingAgentSessionStatus::TimedOut => "timed_out",
-        ExternalCodingAgentSessionStatus::Closed => "closed",
-    }
-}
-
-fn external_coding_agent_session_json(snapshot: &ExternalCodingAgentSessionSnapshot) -> Value {
-    json!({
-        "session_id": snapshot.session_id,
-        "workspace_id": snapshot.workspace_id,
-        "status": external_coding_agent_status_label(snapshot.status),
-        "started_unix_ms": snapshot.started_unix_ms,
-        "last_activity_unix_ms": snapshot.last_activity_unix_ms,
-        "queued_followups": snapshot.queued_followups,
-    })
-}
-
-fn external_coding_agent_event_json(
-    event: &tau_runtime::ExternalCodingAgentProgressEvent,
-) -> Value {
-    json!({
-        "sequence_id": event.sequence_id,
-        "event_type": event.event_type,
-        "message": event.message,
-        "timestamp_unix_ms": event.timestamp_unix_ms,
     })
 }
